Functionnal game
Adding the condition to have only one symbol per box.
Adding a restart button to clear all the symbol
Adding a function to check winning condition
Added a test to see which player have to play but the text is superposing so we cannot see, this need to be fixed
Need to add a line when a player win
Need to adjust the code when the size of the window is different.